🐛 fix: 저장되는 좌표 타입 double로 변경

See merge request s10-final/S10P31A202!85

+++ server/src/main/java/com/zzuiksa/server/domain/place/entity/Place.java
@@ -46,11 +46,11 @@ public class Place extends BaseEntity {
 
     @NotNull
     @Column(nullable = false)
-    private Float lat;
+    private Double lat;
 
     @NotNull
     @Column(nullable = false)
-    private Float lng;
+    private Double lng;
 
     public void setName(String name) {
         if (!Utils.hasTextAndLengthBetween(name, 1, 100)) {
@@ -59,7 +59,7 @@ public class Place extends BaseEntity {
         this.name = name;
     }
 
-    public void setLatLng(@NotNull Float lat, @NotNull Float lng) {
+    public void setLatLng(@NotNull Double lat, @NotNull Double lng) {
         this.lat = lat;
         this.lng = lng;
     }
